requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.5
django-cors-headers==3.13.0
django-oauth-toolkit==2.2.0
django-seed==0.3.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
Faker==16.3.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
Pillow==9.4.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.7
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toposort==1.8
tzdata==2022.7
urllib3==1.26.14
wrapt==1.14.1
